requirements.txt
streamlit
torch
transformers
sentence-transformers
nltk
faiss-cpu
scikit-learn
